# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API LIVROS",
    "description": "Documentação da API de Livros",
    "version": "1.0.0"
  },
  "paths": {
    "/api/livro": {
      "get": {
        "tags": ["Livros"],
        "summary": "Listar todos os livros",
        "responses": {
          "200": {
            "description": "Lista de livros"
          }
        }
      },
      "post": {
        "tags": ["Livros"],
        "summary": "Cadastrar livro",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "titulo": { "type": "string" },
                  "autor": { "type": "string" },
                  "genero": { "type": "string" },
                  "ano": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Livro cadastrado"
          }
        }
      }
    },
    "/api/livro/{id}": {
      "get": {
        "tags": ["Livros"],
        "summary": "Buscar livro por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Livro encontrado"
          },
          "404": {
            "description": "Livro não encontrado"
          }
        }
      },
      "put": {
        "tags": ["Livros"],
        "summary": "Atualizar livro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "titulo": { "type": "string" },
                  "autor": { "type": "string" },
                  "genero": { "type": "string" },
                  "ano": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Livro atualizado"
          },
          "404": {
            "description": "Livro não encontrado"
          }
        }
      },
      "delete": {
        "tags": ["Livros"],
        "summary": "Excluir livro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Livro removido"
          },
          "404": {
            "description": "Livro não encontrado"
          }
        }
      }
    }
  }
}